Moved image_processing.rst one folder up.
Not sure what was originally intended here, but given the way
the links in scipy.rst and image_processing.rst are
structured, it seems like image_processing.rst was intended to
be in the intro/ directory.

Moving the file gets rid of quite a few sphinx warnings.

Two additional links modified to reflect new location of
image_processing.rst.